requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-SocketIO==5.3.4
python-socketio==5.9.0
python-engineio==4.7.1
python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7
PyJWT==2.8.0
pytz==2023.3
gunicorn==21.2.0
eventlet==0.36.1